refactor: the mutation tool was three quarters of another repository's
nanobench's mutate.py was a fork of this one taken before it grew up: 777 of
its 1034 lines were line-identical with this file, and the ~1200 lines added
here since -- the deletions and bitwise operators, the memory cap and the oom
verdict, dropping mutants the preprocessor removes, the pre-filter read out of
compile_commands.json, --diff from the merge base -- were invisible over there.
Two copies of a tool whose every past mistake flattered the tests, and only one
of them with a test suite.

So the part that is not about any one project moves into mutate_core.py, which
nanobench now holds a byte-identical copy of, and mutate.py becomes this
project's adapter: the header, that meson configures the build, that the binary
is udm-test, that a lane wants FUZZ_CORPUS_BASE_DIR, and the constants behind
--dry-run that only a measurement here can supply. ~1800 lines shared against
~100 of adapter.

Both build backends live in the core rather than one in each adapter, which is
the whole point of the arrangement: test_mutate.py covers the cmake backend
nanobench drives, from a repository that builds with meson and will never run
it. A backend covered only where it is used is exactly as untested as it was
before.

Two behaviours arrive with the merge, both from tools that had learned
something this one had not:

- a bug block whose replacement still contains the text it replaced is refused.
  It changes nothing, so `caught` and `SURVIVED` are both verdicts about
  nothing, and it is how a *move* gets written wrongly. woswoar's tools/mutate.py
  shipped three in one session before it started refusing them. `<<< additive`
  is the way to mean it; nanobench has exactly one block that does.
- --exclude-filter, doctest's -tce=, for the machine where one case is flaky.
  The baseline refuses to score until the suite is green twice and that is the
  point of it, so naming the flake beats lowering --baseline-runs, and the
  fingerprint then prints what was skipped.

lint-mutate-core.py fails if this copy has been edited without its recorded
hash moving with it, which is the failure vendoring introduces: a convenient
local fix here leaves the other repository running something nothing tests.
No lint in either repository can see the other one, so the two .sha256 files
are what makes "are they in sync?" a question diff can answer.

Verified: 192 hermetic tests green (18 new, covering the backends, the project
seam, the root-only ignore patterns and the additive rule); all five linters
green; and one real bug from bugs/erase-path.txt run end to end through meson,
which reports the same verdict as the pre-refactor tool did on the same bug --
1 mutant, 597 cases green twice, `compiler` both times.
